changed names of RI scripts and moved old scripts into archive folder
