+++ 04c_validate_coordinated_multi_market.py
@@ -14,7 +14,7 @@ from src.coordinated_multi_market.custom_ppo import CustomPPO
 # from src.coordinated_multi_market.rolling_intrinsic.new_testing_rolling_intrinsic_qh_intelligent_stacking import (
 #     simulate_days_stacked_quarterhourly_products,)
 
-from src.coordinated_multi_market.rolling_intrinsic.testing_rolling_intrinsic_gurobi import (
+from src.coordinated_multi_market.rolling_intrinsic.testing_rolling_intrinsic_qh_intelligent_stacking import (
     simulate_days_stacked_quarterhourly_products,
 )
