+++ src/Archive/new_testing_rolling_intrinsic_qh_intelligent_stacking.py
