+++ src/Archive/rolling_intrinsic_new.py
