+++ src/Archive/rolling_intrinsic_new_copy.py
