+++ src/coordinated_multi_market/custom_ppo.py
@@ -19,7 +19,7 @@ from stable_baselines3.ppo import PPO
 #from src.coordinated_multi_market.rolling_intrinsic.new_training_rolling_intrinsic_qh_intelligent_stacking import (
 #    simulate_days_stacked_quarterhourly_products,)
 
-from src.coordinated_multi_market.rolling_intrinsic.training_rolling_intrinsic_gurobi import (
+from src.coordinated_multi_market.rolling_intrinsic.training_rolling_intrinsic_qh_intelligent_stacking import (
     simulate_days_stacked_quarterhourly_products,)
