+++ src/single_market/rolling_intrinsic_gurobi_qh.py
